+++ std_blocks/pid/pid.c
@@ -8,8 +8,9 @@
 
 #include "pid.h"
 
-def_read_dynarr_fun(read_doubles, double)
-def_write_dynarr_fun(write_doubles, double)
+/* def_read_dynarr_fun(read_doubles, double); */
+def_port_writers(write_double, double)
+def_port_readers(read_double, double)
 
 /* block local state */
 struct pid_info
@@ -142,8 +143,8 @@ void pid_step(ubx_block_t *b)
 	long len_msr, len_des;
 	struct pid_info *inf = (struct pid_info*) b->private_data;
 
-	len_msr = read_doubles(inf->ports.msr, inf->msr, inf->data_len);
-	len_des = read_doubles(inf->ports.des, inf->des, inf->data_len);
+	len_msr = read_double_array(inf->ports.msr, inf->msr, inf->data_len);
+	len_des = read_double_array(inf->ports.des, inf->des, inf->data_len);
 
 	if (len_msr < 0) {
 		ubx_err(b, "ENODATA: no data on port msr");
@@ -190,7 +191,7 @@ void pid_step(ubx_block_t *b)
 		}
 	}
 
-	write_doubles(inf->ports.out, inf->out, inf->data_len);
+	write_double_array(inf->ports.out, inf->out, inf->data_len);
 
 	/* swap ptrs */
 	tmp = inf->err_prev;